update: removed cart controller and routes
