+++ services/product/src/index.ts
@@ -6,7 +6,6 @@ import logger from "./utils/logger";
 import httpLogger from "./middleware/httpLogger";
 import rateLimiter from "./middleware/rateLimitter";
 import cartRoute from "./routes/cart.route";
-import cartItemRoute from "./routes/cartItem.routes";
 
 dotenv.config();
 
@@ -19,7 +18,6 @@ app.use(rateLimiter);
 
 app.use("/product", productRoute);
 app.use("/cart", cartRoute);
-app.use("/cart", cartItemRoute);
 
 app.listen(PORT, async () => {
   await connectDB();
