+++ services/product/src/routes/cart.route.ts
@@ -3,17 +3,21 @@ import {
   createCart,
   deleteCart,
   getCartById,
+  removeCartItem,
   updateCart,
+  upsertCartItem,
 } from "../controllers/cart.controller";
 
 const router = Router();
 
-router.get("/:id", getCartById);
+router.get("/:cartId", getCartById);
 
 router.post("/", createCart);
 
-router.patch("/:id", updateCart);
+router.patch("/:cartId", updateCart);
+router.patch("/:cartId/item", upsertCartItem);
 
-router.delete("/:id", deleteCart);
+router.delete("/:cartId", deleteCart);
+router.delete("/:cartId/item", removeCartItem);
 
 export default router;
