--- services/product/src/routes/cartItem.routes.ts
@@ -1,13 +1,0 @@
-import { Router } from "express";
-import {
-  removeCartItem,
-  upsertCartItem,
-} from "../controllers/cartItem.controller";
-
-const router = Router();
-
-router.patch("/:cartId/items", upsertCartItem);
-
-router.delete("/:cartId/items", removeCartItem);
-
-export default router;
